+++ demo/src/app/app.module.ts
@@ -6,6 +6,7 @@ import { ListItemComponent } from './lists/list-item.component';
 import { ListWithApiComponent } from './lists/list-with-api.component';
 import { MultiColListComponent } from './lists/multi-col-list.component';
 import { NgModule } from '@angular/core';
+import { TableListComponent } from './lists/table-list.component';
 import { VerticalListComponent } from './lists/vertical-list.component';
 import { VirtualScrollModule } from '../../../src/virtual-scroll';
 import { VirtualArrayComponent } from './lists/virtual-array.component';
@@ -16,9 +17,8 @@ import { VirtualArrayComponent } from './lists/virtual-array.component';
         ListItemComponent,
         ListWithApiComponent,
         MultiColListComponent,
-        VerticalListComponent,
-        VirtualArrayComponent
-    ],
+    TableListComponent,    VerticalListComponent,
+    VirtualArrayComponent],
     imports: [
         BrowserModule,
         FormsModule,