+++ demo/src/app/lists/multi-col-list.component.ts
@@ -6,25 +6,28 @@ import { ListItem } from './list-item.component';
   selector: 'multi-col-list',
   template: `
 
-        <button (click)="reduceList()">Reduce to 100 Items</button>
-        <button (click)="setToFullList()">Revert to 1000 Items</button>
+    <button (click)="sortByName()">Sort By Name</button>
+    <button (click)="sortByIndex()">Sort By Index</button>
+    <button (click)="reduceListToEmpty()">Reduce to 0 Items</button>
+    <button (click)="reduceList()">Reduce to 100 Items</button>
+    <button (click)="setToFullList()">Revert to 1000 Items</button>
 
-        <div class="status">
-             Showing <span class="badge">{{indices?.start + 1}}</span>
-             - <span class="badge">{{indices?.end}}</span>
-             of <span class="badge">{{filteredList?.length}}</span>
-            <span>({{scrollItems?.length}} nodes)</span>
-            </div>
+    <div class="status">
+        Showing <span class="badge">{{indices?.start + 1}}</span>
+        - <span class="badge">{{indices?.end}}</span>
+        of <span class="badge">{{filteredList?.length}}</span>
+      <span>({{scrollItems?.length}} nodes)</span>
+      </div>
 
-        <virtual-scroll
-            [items]="filteredList"
-            (update)="scrollItems = $event"
-            (change)="indices = $event">
+    <div virtualScroll
+      [items]="filteredList"
+      (update)="scrollItems = $event"
+      (change)="indices = $event">
 
-            <list-item *ngFor="let item of scrollItems" [item]="item"> </list-item>
+      <list-item *ngFor="let item of scrollItems" [item]="item"> </list-item>
 
-        </virtual-scroll>
-    `,
+    </div>
+  `,
   styleUrls: ['./multi-col-list.scss']
 })
 export class MultiColListComponent implements OnChanges {
@@ -36,8 +39,20 @@ export class MultiColListComponent implements OnChanges {
 
   filteredList: ListItem[];
 
+  reduceListToEmpty() {
+    this.filteredList = [];
+  }
+
   reduceList() {
     this.filteredList = (this.items || []).slice(0, 100);
+  }
+
+  sortByName() {
+    this.filteredList = [].concat(this.filteredList || []).sort((a, b) => -(a.name < b.name) || +(a.name !== b.name));
+  }
+
+  sortByIndex() {
+    this.filteredList = [].concat(this.filteredList || []).sort((a, b) => -(a.index < b.index) || +(a.index !== b.index));
   }
 
   setToFullList() {